Final task commit
Transaction db moved under database folder. All the paths updated accordingly.
Load database function created to use appropriate file to load the transactional database,
